pyproject.toml
[project]
name = "tenorbot"
version = "0.1.0"
description = ""
authors = [{ name = "Yan", email = "me@octonezd.me" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["python-telegram-bot (>=21.11.1,<22.0.0)", "requests (>=2.32.3,<3.0.0)", "aiohttp (>=3.11.13,<4.0.0)", "python-dotenv (>=1.0.1,<2.0.0)"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
